+++ libs/platform/experience/src/services/layout/plugins/static.ts
